+++ src/test/java/com/zelusik/eatery/unit/domain/review/service/ReviewCommandServiceTest.java
@@ -4,11 +4,9 @@ import com.zelusik.eatery.domain.bookmark.service.BookmarkQueryService;
 import com.zelusik.eatery.domain.member.constant.Gender;
 import com.zelusik.eatery.domain.member.constant.LoginType;
 import com.zelusik.eatery.domain.member.constant.RoleType;
-import com.zelusik.eatery.domain.member.dto.MemberDto;
 import com.zelusik.eatery.domain.member.entity.Member;
 import com.zelusik.eatery.domain.member.service.MemberQueryService;
 import com.zelusik.eatery.domain.place.constant.KakaoCategoryGroupCode;
-import com.zelusik.eatery.domain.place.dto.PlaceDto;
 import com.zelusik.eatery.domain.place.entity.Address;
 import com.zelusik.eatery.domain.place.entity.Place;
 import com.zelusik.eatery.domain.place.entity.PlaceCategory;
@@ -20,10 +18,9 @@ import com.zelusik.eatery.domain.review.dto.ReviewDto;
 import com.zelusik.eatery.domain.review.dto.request.ReviewCreateRequest;
 import com.zelusik.eatery.domain.review.entity.Review;
 import com.zelusik.eatery.domain.review.exception.ReviewDeletePermissionDeniedException;
-import com.zelusik.eatery.domain.review.exception.ReviewNotFoundException;
 import com.zelusik.eatery.domain.review.repository.ReviewRepository;
-import com.zelusik.eatery.domain.review.service.ReviewService;
-import com.zelusik.eatery.domain.review_image.dto.ReviewImageDto;
+import com.zelusik.eatery.domain.review.service.ReviewCommandService;
+import com.zelusik.eatery.domain.review.service.ReviewQueryService;
 import com.zelusik.eatery.domain.review_image.dto.request.ReviewImageCreateRequest;
 import com.zelusik.eatery.domain.review_image.entity.ReviewImage;
 import com.zelusik.eatery.domain.review_image.service.ReviewImageService;
@@ -32,40 +29,34 @@ import com.zelusik.eatery.domain.review_image_menu_tag.dto.request.ReviewMenuTag
 import com.zelusik.eatery.domain.review_image_menu_tag.repository.ReviewImageMenuTagRepository;
 import com.zelusik.eatery.domain.review_keyword.entity.ReviewKeyword;
 import com.zelusik.eatery.domain.review_keyword.repository.ReviewKeywordRepository;
-import com.zelusik.eatery.global.common.constant.EateryConstants;
-import com.zelusik.eatery.global.common.constant.FoodCategoryValue;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.SliceImpl;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
-import static com.zelusik.eatery.domain.review.constant.ReviewEmbedOption.PLACE;
-import static com.zelusik.eatery.domain.review.constant.ReviewEmbedOption.WRITER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.*;
 
-@DisplayName("[Unit] Service - Review")
+@DisplayName("[Unit] Service(Command) - Review")
 @ExtendWith(MockitoExtension.class)
-class ReviewServiceTest {
+class ReviewCommandServiceTest {
 
     @InjectMocks
-    private ReviewService sut;
+    private ReviewCommandService sut;
 
+    @Mock
+    private ReviewQueryService reviewQueryService;
     @Mock
     private ReviewImageService reviewImageService;
     @Mock
@@ -123,105 +114,6 @@ class ReviewServiceTest {
         assertThat(actualSavedReview.getPlace().getKakaoPid()).isEqualTo(kakaoPid);
     }
 
-    @DisplayName("리뷰의 id(PK)가 주어지고, id로 리뷰를 단건 조회하면, 조회된 리뷰가 반환된다.")
-    @Test
-    void givenReviewId_whenFindReviewById_thenReturnReview() {
-        // given
-        long memberId = 1L;
-        long reviewId = 2L;
-        Review expectedResult = createReview(reviewId, createMember(memberId), createPlace(3L, "12345"));
-        given(reviewRepository.findByIdAndDeletedAtNull(reviewId)).willReturn(Optional.of(expectedResult));
-
-        // when
-        Review actualResult = sut.findById(reviewId);
-
-        // then
-        then(reviewRepository).should().findByIdAndDeletedAtNull(reviewId);
-        verifyEveryMocksShouldHaveNoMoreInteractions();
-        assertThat(actualResult)
-                .hasFieldOrPropertyWithValue("id", reviewId)
-                .hasFieldOrPropertyWithValue("writer.id", memberId);
-    }
-
-    @DisplayName("존재하지 않는 리뷰 id(PK)가 주어지고, 주어진 id로 리뷰를 단건 조회하면, 예외가 발생한다..")
-    @Test
-    void givenNotExistentReviewId_whenFindReviewById_thenThrowReviewNotFoundException() {
-        // given
-        long reviewId = 2L;
-        given(reviewRepository.findByIdAndDeletedAtNull(reviewId)).willReturn(Optional.empty());
-
-        // when
-        Throwable t = catchThrowable(() -> sut.findById(reviewId));
-
-        // then
-        then(reviewRepository).should().findByIdAndDeletedAtNull(reviewId);
-        verifyEveryMocksShouldHaveNoMoreInteractions();
-        assertThat(t).isInstanceOf(ReviewNotFoundException.class);
-    }
-
-    @DisplayName("리뷰의 id(PK)가 주어지고, id로 리뷰 dto를 단건 조회하면, 조회된 리뷰의 dto가 반환된다.")
-    @Test
-    void givenReviewId_whenFindReviewDtoById_thenReturnReviewDto() {
-        // given
-        long memberId = 1L;
-        long reviewId = 2L;
-        Review expectedResult = createReview(reviewId, createMember(memberId), createPlace(3L, "12345"));
-        given(reviewRepository.findByIdAndDeletedAtNull(reviewId)).willReturn(Optional.of(expectedResult));
-        given(bookmarkQueryService.isMarkedPlace(eq(memberId), any(Place.class))).willReturn(false);
-
-        // when
-        ReviewDto actualResult = sut.findDtoById(memberId, reviewId);
-
-        // then
-        then(reviewRepository).should().findByIdAndDeletedAtNull(reviewId);
-        then(bookmarkQueryService).should().isMarkedPlace(eq(memberId), any(Place.class));
-        verifyEveryMocksShouldHaveNoMoreInteractions();
-        assertThat(actualResult)
-                .hasFieldOrPropertyWithValue("id", reviewId)
-                .hasFieldOrPropertyWithValue("writer.id", memberId);
-    }
-
-    @DisplayName("리뷰 목록을 조회하면, 조회된 리뷰 목록(Slice)을 반환한다.")
-    @Test
-    void given_whenFindDtos_thenReturnReviews() {
-        // given
-        long loginMemberId = 1L;
-        Pageable pageable = Pageable.ofSize(15);
-        MemberDto member = createMemberDto(3L);
-        PlaceDto place = createPlaceDto(4L, "12345");
-        Slice<ReviewDto> expectedSearchResult = new SliceImpl<>(List.of(createReviewDto(2L, member, place)));
-        given(reviewRepository.findDtos(loginMemberId, null, null, List.of(WRITER, PLACE), pageable)).willReturn(expectedSearchResult);
-
-        // when
-        Slice<ReviewDto> actualSearchResult = sut.findDtos(loginMemberId, null, null, List.of(WRITER, PLACE), pageable);
-
-        // then
-        then(reviewRepository).should().findDtos(loginMemberId, null, null, List.of(WRITER, PLACE), pageable);
-        verifyEveryMocksShouldHaveNoMoreInteractions();
-        assertThat(actualSearchResult).hasSize(expectedSearchResult.getSize());
-    }
-
-    @DisplayName("리뷰 피드를 조회한다.")
-    @Test
-    void given_whenFindReviewFeed_thenReturnResults() {
-        // given
-        long loginMemberId = 1L;
-        long reviewId = 2L;
-        Pageable pageable = Pageable.ofSize(10);
-        MemberDto writer = createMemberDto(3L);
-        PlaceDto place = createPlaceDto(4L, "123");
-        Slice<ReviewDto> expectedResults = new SliceImpl<>(List.of(createReviewDto(reviewId, writer, place)), pageable, false);
-        given(reviewRepository.findReviewFeed(loginMemberId, pageable)).willReturn(expectedResults);
-
-        // when
-        Slice<ReviewDto> actualResults = sut.findReviewReed(loginMemberId, Pageable.ofSize(10));
-
-        // then
-        then(reviewRepository).should().findReviewFeed(loginMemberId, pageable);
-        verifyEveryMocksShouldHaveNoMoreInteractions();
-        assertThat(actualResults).hasSize(expectedResults.getNumberOfElements());
-    }
-
     @DisplayName("리뷰를 삭제하면, 리뷰와 모든 리뷰 이미지들을 soft delete하고 모든 리뷰 키워드들을 삭제한다.")
     @Test
     void given_whenSoftDeleteReview_thenSoftDeleteReviewAndAllReviewImagesAndDeleteAllReviewKeywords() {
@@ -236,7 +128,7 @@ class ReviewServiceTest {
         ReviewImage reviewImage = createReviewImage(6L, findReview);
         findReview.getReviewImages().add(reviewImage);
         given(memberQueryService.findById(memberId)).willReturn(loginMember);
-        given(reviewRepository.findByIdAndDeletedAtNull(reviewId)).willReturn(Optional.of(findReview));
+        given(reviewQueryService.findById(reviewId)).willReturn(findReview);
         willDoNothing().given(reviewImageService).softDeleteAll(findReview.getReviewImages());
         willDoNothing().given(reviewKeywordRepository).deleteAll(findReview.getKeywords());
         willDoNothing().given(reviewRepository).flush();
@@ -247,7 +139,7 @@ class ReviewServiceTest {
 
         // then
         then(memberQueryService).should().findById(memberId);
-        then(reviewRepository).should().findByIdAndDeletedAtNull(reviewId);
+        then(reviewQueryService).should().findById(reviewId);
         then(reviewImageService).should().softDeleteAll(findReview.getReviewImages());
         then(reviewKeywordRepository).should().deleteAll(findReview.getKeywords());
         then(reviewRepository).should().flush();
@@ -271,25 +163,28 @@ class ReviewServiceTest {
         ReviewImage reviewImage = createReviewImage(6L, findReview);
         findReview.getReviewImages().add(reviewImage);
         given(memberQueryService.findById(loginMemberId)).willReturn(loginMember);
-        given(reviewRepository.findByIdAndDeletedAtNull(reviewId)).willReturn(Optional.of(findReview));
+        given(reviewQueryService.findById(reviewId)).willReturn(findReview);
 
         // when
         Throwable t = catchThrowable(() -> sut.delete(loginMemberId, reviewId));
 
         // then
         then(memberQueryService).should().findById(loginMemberId);
-        then(reviewRepository).should().findByIdAndDeletedAtNull(reviewId);
+        then(reviewQueryService).should().findById(reviewId);
         verifyEveryMocksShouldHaveNoMoreInteractions();
         assertThat(t).isInstanceOf(ReviewDeletePermissionDeniedException.class);
     }
 
     private void verifyEveryMocksShouldHaveNoMoreInteractions() {
+        then(reviewQueryService).shouldHaveNoMoreInteractions();
         then(reviewImageService).shouldHaveNoMoreInteractions();
         then(memberQueryService).shouldHaveNoMoreInteractions();
+        then(placeCommandService).shouldHaveNoMoreInteractions();
         then(placeQueryService).shouldHaveNoMoreInteractions();
         then(reviewRepository).shouldHaveNoMoreInteractions();
         then(reviewKeywordRepository).shouldHaveNoMoreInteractions();
         then(bookmarkQueryService).shouldHaveNoMoreInteractions();
+        then(reviewImageMenuTagRepository).shouldHaveNoMoreInteractions();
     }
 
     private Member createMember(Long memberId) {
@@ -315,28 +210,6 @@ class ReviewServiceTest {
         );
     }
 
-    private MemberDto createMemberDto(Long memberId) {
-        return createMemberDto(memberId, Set.of(RoleType.USER));
-    }
-
-    private MemberDto createMemberDto(Long memberId, Set<RoleType> roleTypes) {
-        return new MemberDto(
-                memberId,
-                EateryConstants.defaultProfileImageUrl,
-                EateryConstants.defaultProfileThumbnailImageUrl,
-                "1234567890",
-                LoginType.KAKAO,
-                roleTypes,
-                "test@test.com",
-                "test",
-                LocalDate.of(1998, 1, 5),
-                20,
-                Gender.MALE,
-                List.of(FoodCategoryValue.KOREAN),
-                null
-        );
-    }
-
     private Place createPlace(long id, String kakaoPid) {
         return Place.of(
                 id,
@@ -356,26 +229,6 @@ class ReviewServiceTest {
         );
     }
 
-    public static PlaceDto createPlaceDto(long placeId, String kakaoPid) {
-        return new PlaceDto(
-                placeId,
-                List.of(ReviewKeywordValue.FRESH),
-                kakaoPid,
-                "연남토마 본점",
-                "http://place.map.kakao.com/308342289",
-                KakaoCategoryGroupCode.FD6,
-                PlaceCategory.of("음식점 > 퓨전요리 > 퓨전일식"),
-                "02-332-8064",
-                Address.of("서울 마포구 연남동 568-26", "서울 마포구 월드컵북로6길 61"),
-                "http://place.map.kakao.com/308342289",
-                new Point("37.5595073462493", "126.921462488105"),
-                null,
-                List.of(),
-                null,
-                false
-        );
-    }
-
     private Review createReview(Long reviewId, Member member, Place place) {
         return Review.of(
                 reviewId,
@@ -386,26 +239,6 @@ class ReviewServiceTest {
                 LocalDateTime.now(),
                 LocalDateTime.now(),
                 null
-        );
-    }
-
-    public static ReviewDto createReviewDto(long reviewId, MemberDto writer, PlaceDto place) {
-        return new ReviewDto(
-                reviewId,
-                writer,
-                place,
-                List.of(ReviewKeywordValue.NOISY, ReviewKeywordValue.FRESH),
-                "자동 생성된 내용",
-                "제출된 내용",
-                List.of(new ReviewImageDto(
-                        1L,
-                        1L,
-                        "test.txt",
-                        "storedName",
-                        "url",
-                        "thumbnailStoredName",
-                        "thumbnailUrl")),
-                LocalDateTime.now()
         );
     }
