+++ src/test/resources/junit-platform.properties
@@ -1,0 +1,1 @@
+junit.jupiter.testclass.order.default = org.junit.jupiter.api.ClassOrderer$DisplayName
